+++ bin/uninstall.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+service impress stop
+chkconfig impress off
+chkconfig --del impress
+rm /etc/init.d/impress